# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Nika Web Services",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1"
        }
    ],
    "paths": {
...
        "/roles": {
            "get": {
                "tags": [
                    "نقش‌های کاربری"
                ],
                "summary": "لیست نقش‌های کاربری همراه مجوزهایشان بصورت صفحه بندی",
                "description": "",
                "operationId": "ce32414d6aa6fc86e1ed8a8cce347152",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "شماره صفحه",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "example": 1
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "تعداد نمایش در هر صفحه",
                        "schema": {
                            "type": "number"
                        },
                        "example": 10
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "جستجوی نام انگلیسی و فارسی",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "عملیات موفق",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "نقش‌های کاربری"
                ],
                "summary": "ثبت نقش کاربری",
                "description": "",
                "operationId": "e498c068c0457de5174ea728d6fb03b3",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "name_fa"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "نام انگلیسی",
                                        "type": "string"
                                    },
                                    "name_fa": {
                                        "description": "نام فارسی",
                                        "type": "string"
                                    },
                                    "permissions[]": {
                                        "description": "مجوزها (نام مجوز)",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "ثبت با موفقیت انجام شد",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
        "/roles/{id}": {
            "post": {
                "tags": [
                    "نقش‌های کاربری"
                ],
                "summary": "بروزرسانی نقش کاربری",
                "description": "نام انگلیسی و فارسی نقش کاربری هایی که قفل هستند، قابل تغییر نمیباشند",
                "operationId": "f1284214d9d3e387aabc5cd3e45cc585",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "شناسه نقش کاربری",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "_method"
                                ],
                                "properties": {
                                    "_method": {
                                        "description": "این مقدار باید بصورت ثابت شود",
                                        "type": "string",
                                        "default": "put",
                                        "enum": [
                                            "put"
                                        ]
                                    },
                                    "name": {
                                        "description": "نام انگلیسی",
                                        "type": "string"
                                    },
                                    "name_fa": {
                                        "description": "نام فارسی",
                                        "type": "string"
                                    },
                                    "permissions[]": {
                                        "description": "مجوزها (نام مجوز)",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "عملیات موفق",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "نقش‌های کاربری"
                ],
                "summary": "حذف نقش کاربری",
                "description": "نقش کاربری هایی که قفل هستند، قابل حذف نمیباشند",
                "operationId": "7e347d12d97a404c4084d1d7cc641f42",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "شناسه نقش کاربری",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "عملیات موفق",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "توکن را به همراه Bearer وارد کنید.",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}